test: add reusable steam env helper
