+++ src/cli/config_paths.rs
@@ -32,15 +32,18 @@ pub fn execute() {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_helpers::TEST_MUTEX;
+    use crate::test_helpers::{setup_steam_env, TEST_MUTEX};
     use std::fs;
     use tempfile::tempdir;
 
-    fn setup_mock_userdata() -> (tempfile::TempDir, std::path::PathBuf) {
-        let home = tempdir().unwrap();
+    #[test]
+    fn test_execute_emits_found_paths() {
+        let _guard = TEST_MUTEX.lock().unwrap();
+        crate::core::steam::clear_caches();
+        let (home, _prefix, login_opt) = setup_steam_env(0, true);
+        let login = login_opt.unwrap();
         let userdata = home.path().join(".steam/steam/userdata");
         fs::create_dir_all(&userdata).unwrap();
-
         let u1 = userdata.join("111111111");
         let u2 = userdata.join("222222222");
         fs::create_dir_all(u1.join("config")).unwrap();
@@ -49,24 +52,12 @@ mod tests {
         let cfg2 = u2.join("config/localconfig.vdf");
         fs::write(&cfg1, "").unwrap();
         fs::write(&cfg2, "").unwrap();
-
-        let config_dir = home.path().join(".steam/config");
-        fs::create_dir_all(&config_dir).unwrap();
-        let login = config_dir.join("loginusers.vdf");
         let contents = r#""users" {
             "111111111" { "MostRecent" "0" }
             "222222222" { "MostRecent" "1" }
         }"#;
         fs::write(&login, contents).unwrap();
-
-        (home, cfg2)
-    }
-
-    #[test]
-    fn test_execute_emits_found_paths() {
-        let _guard = TEST_MUTEX.lock().unwrap();
-        crate::core::steam::clear_caches();
-        let (home, expected) = setup_mock_userdata();
+        let expected = cfg2.clone();
         let old_home = std::env::var("HOME").ok();
         std::env::set_var("HOME", home.path());
